requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pyarrow>=14.0.0
ftfy>=6.1.0
langdetect>=1.0.9
rapidfuzz>=3.5.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
pydantic>=2.0.0
streamlit>=1.30.0
lxml>=4.9.0
plotly>=5.18.0
openai>=1.30.0
python-dotenv>=1.0.0
